// types.ts
export enum Page {
  Home = 'Home',
  About = 'About Us',
  Resources = 'Resources',
  Contact = 'Contact',
  Blog = 'Blog',
  Shop = 'Shop',
  Paternity = 'Paternity',
  Custody = 'Custody',
  ChildSupport = 'Child Support',
  Fatherhood = 'Fatherhood',
  // Paternity Sub-pages
  PaternityTesting = 'Paternity Testing',
  PaternityFraud = 'Paternity Fraud',
  LegalDocuments = 'Legal Documents',
  // Fatherhood Sub-pages
  NewDads = 'For New Dads',
  CoParenting = 'Co-Parenting Strategies',
  FathersWellbeing = "Father's Well-being",
  // Child Support Sub-pages
  EnforcingSupport = 'Enforcing Child Support',
  SupportModifications = 'Modifying Child Support',
  // Legal Pages
  DisclaimerPage = 'Disclaimer',
  More = 'More',
}